Game model in database
This model, at the database layer, only uses a single text field for
storing JSON data.

The reason for this is that the data structure for a game is complex
enough (by using arrays) to require multiple tables when stored
relationally in a SQL database, but our application requirements have no
need to consider those fields (such as players) as separate models.

If we could assume that the DBMS in use was Postgres we could use its
native capability to store JSON documents, but for local development we
will be using sqlite3 which does not have this feature (by default).

Instead we'll store the JSON document as a text field and transparently
convert it at the ActiveModel layer.