Refactor Game factory & required attributes tests
This is a pretty big shift, but we're properly taking advantage of
Factory Girl's advanced features now.

The Game model is actually constructed using just the data attribute,
but the individual pieces of that hash are exposed as normal Factory
Girl attributes which gives us significantly more flexibility when
writing more detailed validations tests.